Add a AbstractBlockChain.forceReorgToHash(hash) method.
This is useful when doing reverse-headers sync and a block that
had previously been added to a regular BlockChain was found to be
invalid during the second download pass.

The idea is that an initial download pass will be completed using a
regular BlockChain connected to a FullPrunedBlockStore.  Once that
chain is up-to-date, we will go back and fill in full blocks using
a second FullPrunedBlockChain connected to the same
FullPrunedBlockStore.  During the second phase, wallets can remain
connected to the original BlockChain, allowing everything to
continue to function while the second download phase is in
progress.  Once complete, wallets can be disconnected from the
BlockChain and connected to the FullPrunedBlockChain and the
BlockChain can be garbage collected.
However, if a previously accepted block is found to be invalid, we
will need a way to force a reorg in the BlockChain so that wallets
remain consistent when reconnected to the FullPrunedBlockChain
which is at a different height - hence forceReorgToHash.